Language corrections
I came to see if the code could be adapted to biological research, then noticed the problematic English (e.g. "under way") and references to "NSA" instead of "the NSA". Don't claim the code is open-source if it isn't, and don't assure us about efforts being underway. Just post the code and revise the documentation when you do.