Fix 'window' menu key binding
To avoid duplicate actions with existing 'showaction' toolbar buttons,
these toolbar actions were merged in the showAction of the menu.
To keep compatibility with previous shorcuts, owners of actions in 'window'
were changed to the owner of the provider and the action name can be set in
the ComponentProvider class.