Potentially ambiguous macro operator names used
4 operators are emitted in finally C code: ZEXT, SEXT, CONCAT, SUB
Although generally the sizes are just one character (from 1 to 8) potentially numeric values can be 16 bytes long.
The suffix 112 to any of these for example.  Could be 1 byte and 12 bytes.  Or 11 bytes and 2 bytes.  Obviously the naming convention here is poorly chosen.

However that being said, this behavior could be deliberately depended upon in other areas though AFAIK only custom tools would do so.  It is better to change it earlier than later partly for that reason as tools could learn to depend on the behavior as opposed to doing what I have done and choosing to fix the ambiguity in the C source to begin with.