VarMap stable_sort throws Invalid Comparator
Can thank Visual Studio 2019 debug build for finding this one.  Only debug build runtime error but this is a legitimate violation of C++ standards.

stable_sort requires a weak strict ordering yet inspection showed duplicates were ending up in the maplist.

The solution is complexity-wise not ideal.  But it certainly works to prevent this bug.