Segmented Constant Ptrs fail to resolve
Constant pointers to the data segment are not resolving because despite isPointer properly resolving them, the outer apply routine simply ignores the address result and creates a PTR_SUB operation based on the wrong original offset.  Just need to take the offset off of the entry object which has already been verified.