Update fspec.cc
FIx code smell: call mechanisms could only all be disabled (-1) as new IDs are always generated with new names so it would be impossible for a protocol resolver with identical uponreturn or uponentry mechanisms to be loaded - though unnecessarily so.  Instead if all the properties are identical then allow it - perhaps with a warning to the user.

If both are dynamic - theoretically they could be different and it could cause issue.
The body tag is not compared and it could be different and cause issue.
But nonetheless the feature should not be disabled so quickly as sometimes its useful to have a resolver with identical uponreturn/uponentry which currently is impossible.  The body is not parsed or saved but it could be if it were really an issue.  Dynamic ones would simply need to be designed to be the same or should not be used.

I currently solve CS register tracking bugs with an uponreturn dynamic injection on __stdcall16far but I want a resolver with __stdcall16near.  A real legitimate use case scenario which is why I found this issue in the first place.