Bad ranges from client cause a crash
This addresses ranges where the last is less than first.  The range is never validated and it would cause a crash in all sorts of places specifically in my case for ScopeGhidra::processHole when it calls insertRange causing problems with the iterator.  But likely the same issue would occur nearly anywhere its not validated based on the assumption of the coding style.