Infinite loop on jump tables in 64 bit addressing
Jump tables with a NZMask of 0xFFFF000000000000 would cause the int4 stride value to overflow to 0 on shift left after 32 times.  At that point a 0 stride is returned.  In the range this then causes a modulo division by 0 which VS2019 caught immediately.  However in the current code somehow the division by 0 is not thrown and instead it just causes an infinite querying loop of the same address.