Comment out of bound causes decompiler crash
If the client considers the function EOL comments to be in ranges the decompiler does not have a basic block defined for such as a jump table, etc, then the decompiler crashes because it fails to do a null pointer check.  The solution is simple and included.

Tested and working.