Locked arg types cause crash wth unlocked model
This problem looks to be rather deeply ingrained in the model.  However, just because the Java client follows this pattern strictly hardly makes it correct behavior.

If 1 argument is specified and its type is known so it is locked (perhaps the argument is passed itself as an argument to a well known function for example).  The model for the function however is not known and not locked.  Then prototype analysis will NULL out the first argument while detecting additional arguments in slots 2-...  This sets up a good variety of crash places.  I am fixing a few that I specifically found crashed and will continue fixing them as I find them since not all getParam() places are possible when calls like isInputLocked return false.