Decompiler crash with call injection
This seems like a common bug in the decompiler.  CPUI_CALL/IND do not guarantee the first argument is an FSPEC.  Some places check before trying to assume it, while others do not and can crash.  I ran into this bug when I applied my own fix in constant pointers and also forget to check this exact case.  I do not know exactly why some calls do not have this and when but its certainly a normal possibility in the decompiler.

Probably a small audit of getFspecFromConst and this fix applied can reduce decompiler crashes significantly.