More than 20 operands crashes sleigh parser
I have an x86-16 opcode starting with the 16 byte stream 0x00000000100b1437:
0x66 0x0f 0xfc 0xc2 0x66 0x0f 0x6f 0xd0 0xf3 0x0f 0x6f 0x44 0x02 0x10 0x48 0x83

sleigh.cc initializes the operands to a fixed value of 20: DisassemblyCache::initialize -> pos->initialize(75,20,constspace);

Yet this should be dynamically resized in appropriate cases as this byte sequence led to a 21st operand and a crash.  allocateOperand should be a sure place as its the only setter before all the various getters.