Add Visual Studio 6.0 and Windows 98 SDK .prf/.gdt
Add a .prf that can handle the Visual Studio 6.0 and
Windows 98 SDKs.  This is useful for older applications
that were developed pre-NT and don't quite fit with the modern
Windows SDK headers in VS9 and onwards.

Also add a .gdt which was generated directly from this .prf.
Includes DirectSound, DirectDraw, and Direct3D headers.

Signed-off-by: Shane Peelar <lookatyouhacker@gmail.com>