GT-3521 fix problem with "Method is not Remote" when connecting
A recent change in java's RemoteObjectInvocationHandler added an
additional check to ensure that the interface a method was declared on
was also marked with the Remote marker interface.  Previously the check
was looser and only checked if the entire proxied object implemented
Remote.

The change was made in commit be35f9ef53774a87662ad7a0bb978986ea56ca78:

https://github.com/openjdk/jdk/commit/be35f9ef53774a87662ad7a0bb978986ea56ca78

The error the user would encounter would say something about "Method is
not Remote" and something about "RepositoryServerHandle.isReadOnly()".