Add missing semi-colon for function prototypes in headers
When the CREATE_HEADER_FILE is "true" in the CppExporter, function
prototypes get written to the exported header, but the prototypes lack
the semi-colon, and thus fails a syntax check. This fixes that glitch,
by adding the semi-colon during the function prototype export. The code
should be unchanged for the .c export file, in either case.