GP-31 - fix Ghidra File chooser rendering on high dpi displays
Filename display in the directory list mode were getting incorrectly
truncated and shown with ellipses on high DPI display mode.

The previous method was to manually calculate the necessary width of the
directory list's cell renderer, but for some reason we were getting
slightly different values when measuring strings when doing it in our
code vs when swing's JLabel was painting, possible causing our manually
sized cell renderer to be a few pixels too small.

This fix is to yank all that out and let swing measure each element of
the list to determine the necessary width.

Also fixes the directory list cell renderer to inherit the
-Dfont.size.override=XX setting from its parent panel instead of
reseting the font back to non-overidden default.