GP-154 Fix FS refhandle leakage, and loadspec stream stuff
Opening the same container file as a filesystem twice would result in
the second filesystem overwriting the cache entry for the first
filesystem.

Also fix how local files are accessed so that their md5 checksum is
preserved better, to prevent multiple recalculations.

Fix the CsHintLoadSpecChooser filter stream to not die when a loader
returns a null value.