simplify div1 logic
The logic table for this is actually more condensed than the instruction summary makes it seem. This also fixes a bug in the previous code where $(Q_FLAG) was set incorrectly for the old_q = 1, M = 0, Q = 1 case.

Unfortunately I don't have a build setup for ghidra yet so I'm just verifying with my noggin. I will try to verify on a proper build soon.