move Java block comment support and add tests
Move the functionality to handle Java block comments out of the base
`GhidraScriptProvider` class into the `JavaScriptProvider` subclass.
Default behavior is now to not support block comments and rely on
extensions to implement these themselves.

Add a number of tests for Java and Python `ScriptInfo` generation
to ensure that scripts with no block comments, single-line block
comments, and multi-line block comments are all handled appropriately.