Fix manual switching in the middle of an .idx file
Several .idx files in the distribution, such as PowerISA.idx, switch manuals in the middle with a description. Currently, the incorrect manual is used for instructions after the switch. With this fix, switching manuals in the middle of the .idx file works as expected.

This also makes it match the behaviour of the description-less switch at line 1349.