Updating Java Documentation in InstallationGuide.html
This adds a mention of installing Java through package manager, which
might not need modification of path. Also, it makes more obvious the
choice to place the path to Java in the system path, something that I
have noticed is often mistaken. Lastly updated the Windows Java Path
setup to include the new style of path modification dialog.