Created an install sh file
I have created a download / install file for Linux users to ease and ultimate automate the process of installation of Ghidra on the system.
It's my first fully written sh file and I thing it would be wonderful if a professional reviewed it once(the condition statement). The script has an option to download custom version of Ghidra (for those who wish to download an older version) or the default is the latest version. However this program needs to be updated with the link of the latest version.