Use short circuit operators when possible
There are places were a bitwise operator is accidentally used instead of a logical one. In most of these cases, the effect is benign, just costing a tiny bit of performance. However, in cases like (trace != null) & trace... this can cause crashes where it isn't necessary, especially when the first condition is meant to prevent the second condition from happening if a variable/object was null.

This PR fixes those mistakes.