WebAssembly: Emulation support
This patch implements a EmulateInstructionStateModifier for Wasm which provides
support for emulating Wasm instructions.

The memory contents must contain a full module for this to work, because
instruction semantics still depend on module details (e.g. the type of certain
operations depends on metadata like the types of imports or globals).