GDB: Fix parsing of \e and \a
The GDB "printchar" function encodes \x1b as \e and \x07 as \a, both of which
are not handled by StringEscapeUtils.unescapeJava. Strings that contain these
escape characters were incorrectly decoded as "e" and "a" respectively.

Switch to a simple decoding loop instead. The cases here match precisely with
the escape characters used in GDB's printchar function - in particular, there
are no hex escapes, and all octal escapes are exactly three digits long.