x86-64-gcc: Support passing structs in regs, per SysV.
The AMD64 SysV ABI allows most structures up to two QWORDs in size to be passed
in a pair of registers instead of on the stack. This patch handles the case(s)
where the fields are all INTEGER (int or pointer) types.

In the event that a struct has floating-point members, those members are
allocated to the XMM registers if possible. A struct may contain both integral
types as well as floating-point types, meaning that a 16-byte struct may
feasibly be split into one GPR and one SSE register. `metatype=float` does not
examine structure fields, however, so structs with floating-point fields are not
correctly allocated by this patch.