Assume developmentMode if -DbinaryPath is present
`-DbinaryPath=...` controls the `utility.module.ModuleUtilities#BINARY_PATH`
variable which is only used in
`utility.module.ModuleUtilities#getModuleBinDirectories`, which is only
called as part of `ghidra.GhidraLauncher#addModuleBinPaths`, which is only
called if `SystemUtilities.isInDevelopmentMode()` returns true.

This is a fairly obscure command line argument that isn't
documented and only useful for specifing different paths for module
class files when running a dev setup with no IDE or an IDE other than
Eclipse.
Thus it can be safely assumed that if it is present then the
user really wants to run under developmentMode, and doesn't want to rely  on
other checks for this to work.