decompiler(1/4): Add 'using namespace std' to all '.cc' files
This commit will not compile.

It is less controversial (but still not ideal) to add 'using namespace'
to C++ source files because they are isolated and will not affect any
other code than what's in the source file. Eventually, it would be best
to refactor to either use a using-declaration statement or to attach
'std::' to all the symbols. A Clangd automated refactor[1] exists to do
this, but it doesn't work on header files.

I modified the *.y and *.l files, but I did not regenerate because it
would cause more diffs than necessary, but if you did regenerate, it
would generate the intended modifications made to the generated files.

[1]: https://github.com/llvm/llvm-project/blob/llvmorg-16.0.0/clang-tools-extra/clangd/refactor/tweaks/RemoveUsingNamespace.cpp