Add type selection fallback for fs open.
Right now it is not possible to either force or to pick a filesystem
provider that does not have content probing facilities.  This patch adds
a fallback filesystem provider selection choice screen populated with
filesystems that are available but do not probe their contents if no
probing filesystem claimed the resource in question.