spectest bugfix: fix parsing of `select funcref`
Parsing of `select funcref` (1C 01 70) was failing because SLEIGH was committing
to one of the select forms (in this case, `select valtype_64`) using a
rudimentary mask check before going on to fully check the final byte. The LEB128
parsing mixed with the `bytev` checking resulted in a very simplistic mask check
for `select valtype_64`, which passed for 0x70, and resulted in a parsing
failure when `0x70` later failed to match `0x7e` or `0x7c`. The
DebugSleighInstructionParse script was instrumental in identifying this
behaviour.

Instead, the better solution is to check for the 0x01 LEB128 count as early as
possible, and only then examine the final type byte.

This change also improves the display of "ref.null t" and "select t*"
instructions by using symbolic type names.