Fix Mach-O import when no segment contains the header
Certain Mach-O files don't have a segment that contains the header.
In this case, `MachoProgramBuilder.setupHeaderAddr` is supposed to
create an area in the "OTHER" address space for the header.
`setupHeaderAddr` keeps track of the lowest file offset for each
segment that it looks at to find the one with the lowest start
address.  It then adds the "HEADER" memory block as 0 to the lowest
start address.

Before, if there were any segments that were
empty (`segment.getFileOffset() == 0 && segment.getFileSize() == 0`),
then `setupHeaderAddr` would mistakenly set the lowest start address
to 0 instead of ignoring the segment.  This would cause
`MemoryMapDB.checkFileBytesRange` to throw an
`IllegalArgumentException`.

With this change, `setupHeaderAddr` ignores empty segments.

This change passes all of the unit and integration tests.  There are
no explicit tests for `MachoProgramBuilder`.