Add support for HD6301 and HD6303 microcontrollers
HD6303 is a Hitachi clone of 6803. This implementation has been
done based on the Hitachi HD6301V1/HD6303R User's Manual.

The flags are implemented as pseudo-registers for better decompilation
support. The logic for setting the result flags has been optimized to
not produce extra reads from the source address to improve decompilation
of volatile data accesses.