feat: enable match by partial category match
Allow matching when only a partial category match. For example, any
datatype with a namespace below the global namespace such as
`namespace::data` would not resolve since the category was not an exact
match. The new algorithm will search for a match somewhere in the tree.

It currently will not handle multiple matches and will only return the
first match and matches the current behavior for other multi matches.