decompiler: fix INT_LEFT for constant bitshift by 0 and varnodes > 8
Fixes https://github.com/NationalSecurityAgency/ghidra/issues/8705

In case of a 0-bitshift, the mask computation was performing a 64-bit shift,
assuming the result will be 0. However, such a bitshift is undefined and
will be equal to a NOP on some architectures.

This is fixed by explicitly checking for a shift-amount of 0.