pyghidra: Remove some misleading ghidra_builtins entries
I compared the exports declared in the ghidra_builtins stubs against
what was actually available at runtime and I found some names that were
present in the stubs but not available:

    decorate
    decorateOutput
    loadPropertiesFile
    loadVariablesFromState
    updateStateFromVariables
    promptToKeepChangesOnException

As far as I can tell:

* Most of these are protected methods and it's enought to check
  child's kind fo exclude them.
* The above leaves out decorateOutput which is a field but it's not
  explicitly exposed, therefore I copied the explicitly exposed field
  list from elsewhere.

The side effect of the above is the PyGhidraGhidraScript.run method is
also no longer present in the stubs *but* it's still available at
runtime. I think this is fine? I don't see much point in calling it
anyway, all you'll get is an recursion explosion.